Fix RuboCop lint
This commit updates RuboCop configuration to account for changes in
PDK 2.3.0. The biggest functional change is the introduction of
the safe navigation operator, `&.`, into `get_item.rb`. This puts
a hard requirement on using Ruby 2.3 or newer which is satisfied
by Puppet 5.x (used Ruby 2.4) and Bolt 2.x (used Ruby 2.5).