Remove OP_SESSION variable validation
The 1password cli now uses a hashed account identifier when
setting the `OP_SESSION_*` environment variable. This is
a security improvement, but does render the sanity checking
in the `get_item` task useless.

This commit removes the check on the assumption that the
CLI will produce informative error messages if sessions
are not correctly established.